requirements.txt
ipython
nltk-3.2.2
